+++ CHANGELOG.md
@@ -23,6 +23,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Any user can now assign a project to any other user, via the "Assign to" row
   on the project's "Internal contacts" tab.
 - add healthcheck endpoint
+- The form to create a voluntary conversion project allows the user to indicate
+  if the project is being handed over to Regional Casework Services
 
 ## [Release 14][release-14]